feat: Added Google and email/password authorizatoin using Identity Platform
I configured the web app so that users can be authorized with their email and password or by using Google. If the user already has an email in use and is trying to sign in using google, rather than allowing google to override their login, we will ask the user to use their email/password instead. The desigin is very bad right now, but this PR is just meant to show that the Authorization works.