feat: Created onboarding skeleton with file uploading function
I created the onboarding pages without fully implementing the design as it is likely going to be changed soon. However, I still went ahead and added the global fonts Mermaid and DM Sans that will be used throughout the application. I also implemented the file upload functionality with a temporary endpoint.