feature: added timestamp caching. Deleted AppNavbar. Changed Navbar to have Sign Out.
I added timestamps to the cache so that if users are logging in within an hour of their last sent message, they will get their conversations from the cache. I have gotten rid of my Navbar and have used Shaz's Navbar everywhere. I conditionally rend the Login and Sign out button on the navbar depending on whether the user is logged in or not.